[content-store2] Merge the alias providers into the content providers. (#1367)
The previous traits weren't practical for callers, they needed to have both an `AliasProvider` & a `ContentProvider` in hand in some cases, such as for `source_pull()`. We now merge those two into the respective `ContentReader` and `ContentWriter` traits.

Also implemented the alias API in the LocalProvider.

Additionally, provide an aggregator for the AWS S3 provider along with the DynamoS3 provider, in order to provide the alias API (which S3 alone cannot provide).